requirements.txt
asgiref
Django
django-extensions
django-tinymce4-lite
jsmin
Pillow
psycopg2
pydot
pyparsing
sqlparse
tzdata
django-cors-headers
djangorestframework
djangorestframework-simplejwt
django-bootstrap4
gunicorn
django-heroku
whitenoise